.NetFramework版からマージ

+++ Implem.Pleasanter/Libraries/DataTypes/Attachment.cs
@@ -3,6 +3,7 @@ using Implem.Libraries.DataSources.SqlServer;
 using Implem.Libraries.Utilities;
 using Implem.Pleasanter.Libraries.DataSources;
 using Implem.Pleasanter.Libraries.Requests;
+using Implem.Pleasanter.Libraries.Settings;
 using System.Collections.Generic;
 using System.IO;
 using System.Runtime.Serialization;
@@ -19,6 +20,10 @@ namespace Implem.Pleasanter.Libraries.DataTypes
         public bool? Added;
         public bool? Deleted;
         public string Base64;
+        public string FileName;
+        public string Base64Binary;
+        public string HashCode;
+        public bool? Overwritten;
 
         public Attachment()
         {
@@ -27,13 +32,15 @@ namespace Implem.Pleasanter.Libraries.DataTypes
         [OnDeserialized]
         private void OnDeserialized(StreamingContext streamingContext)
         {
-            if (!Base64.IsNullOrEmpty())
+            if (!Base64.IsNullOrEmpty() || !Base64Binary.IsNullOrEmpty())
             {
+                var bin = GetBin();
                 Guid = Strings.NewGuid();
-                Size = GetBin().Length;
-                Extention = Path.GetExtension(Name);
+                Size = bin.Length;
+                Extention = Path.GetExtension(Name ?? FileName);
                 ContentType = Strings.CoalesceEmpty(ContentType, "text/plain");
                 Added = true;
+                Files.Write(bin, Path.Combine(Directories.Temp(), Guid, Name ?? FileName));
             }
         }
 
@@ -63,46 +70,47 @@ namespace Implem.Pleasanter.Libraries.DataTypes
 
         public void WriteToLocal(Context context)
         {
-            if (Added == true)
-            {
-                GetBin().Write(Path.Combine(
+            new FileInfo(Path.Combine(Directories.Temp(), Guid, Name ?? FileName))
+                .CopyTo(Path.Combine(
                     Directories.BinaryStorage(),
                     "Attachments",
-                    Guid));
-            }
-            else if (Deleted == true && !Parameters.BinaryStorage.RestoreLocalFiles)
+                    Guid), overwrite: true);
+        }
+
+        public void DeleteFromLocal(Context context)
+        {
+            var path = Path.Combine(
+                Directories.BinaryStorage(),
+                "Attachments",
+                Guid);
+            if (System.IO.File.Exists(path))
             {
-                Files.DeleteFile(Path.Combine(
-                    Directories.BinaryStorage(),
-                    "Attachments",
-                    Guid));
+                Files.DeleteFile(path);
             }
         }
 
-        public void SqlStatement(Context context, List<SqlStatement> statements, long referenceId)
+        public void SqlStatement(Context context, List<SqlStatement> statements, long referenceId, Column column)
         {
             if (Added == true)
             {
-                var bin = GetBin();
-                if (bin != null)
-                {
-                    statements.Add(Rds.InsertBinaries(
-                        param: Rds.BinariesParam()
-                            .TenantId(context.TenantId)
-                            .ReferenceId(referenceId, _using: referenceId != 0)
-                            .ReferenceId(raw: Def.Sql.Identity, _using: referenceId == 0)
-                            .Guid(Guid)
-                            .Title(Name)
-                            .BinaryType("Attachments")
-                            .Bin(bin, _using: !Parameters.BinaryStorage.IsLocal())
-                            .FileName(Name)
-                            .Extension(Extention)
-                            .Size(Size)
-                            .ContentType(ContentType)));
-                }
+                var bin = IsStoreLocalFolder(column) ? default : GetBin();
+                statements.Add(Rds.UpdateOrInsertBinaries(
+                    param: Rds.BinariesParam()
+                        .TenantId(context.TenantId)
+                        .ReferenceId(referenceId, _using: referenceId != 0)
+                        .ReferenceId(raw: Def.Sql.Identity, _using: referenceId == 0)
+                        .Guid(Guid)
+                        .Title(Name ?? FileName)
+                        .BinaryType("Attachments")
+                        .Bin(bin, _using: !IsStoreLocalFolder(column))
+                        .FileName(Name ?? FileName)
+                        .Extension(Extention)
+                        .Size(Size)
+                        .ContentType(ContentType),
+                    where: Rds.BinariesWhere().Guid(Guid)));
                 DataSources.File.DeleteTemp(Guid);
             }
-            else if (Deleted == true)
+            else if (Deleted == true && !Overwritten.HasValue)
             {
                 statements.Add(Rds.DeleteBinaries(
                     factory: context,
@@ -110,11 +118,36 @@ namespace Implem.Pleasanter.Libraries.DataTypes
             }
         }
 
+        public void SetHashCode(Column column)
+        {
+            if (IsStoreLocalFolder(column))
+            {
+                var filename = Path.Combine(Directories.Temp(), Guid, Name ?? FileName);
+                using (var stream = new FileStream(filename, FileMode.Open, FileAccess.Read))
+                {
+                    var sha = new System.Security.Cryptography.SHA256CryptoServiceProvider();
+                    HashCode = System.Convert.ToBase64String(sha.ComputeHash(stream));
+                }
+            }
+            else
+            {
+                var bytes = GetBin();
+                var sha = new System.Security.Cryptography.SHA256CryptoServiceProvider();
+                HashCode = System.Convert.ToBase64String(sha.ComputeHash(bytes));
+            }
+        }
+
         private byte[] GetBin()
         {
-            return Base64.IsNullOrEmpty()
-                ? Files.Bytes(Path.Combine(Directories.Temp(), Guid, Name))
-                : System.Convert.FromBase64String(Base64);
+            var bin = Base64 ?? Base64Binary;
+            return bin.IsNullOrEmpty()
+                ? Files.Bytes(Path.Combine(Directories.Temp(), Guid, Name ?? FileName))
+                : System.Convert.FromBase64String(bin);
+        }
+
+        public bool IsStoreLocalFolder(Column column)
+        {
+            return Pleasanter.Models.BinaryUtilities.BinaryStorageProvider(column, Size.GetValueOrDefault()) == "LocalFolder";
         }
     }
 }
